refactor: split game_routes.py into domain-specific route modules (#414)
Split the monolithic 2,322-line game_routes.py into 9 focused modules:
- character_routes.py (7 endpoints)
- campaign_routes.py (7 endpoints)
- combat_routes.py (4 endpoints)
- spell_routes.py (7 endpoints)
- dice_routes.py (3 endpoints)
- npc_routes.py (4 endpoints)
- ai_routes.py (4 endpoints)
- session_routes.py (5 endpoints)
- item_routes.py (2 endpoints)

game_routes.py is now a backward-compatible shim that re-exports
from the new modules so existing tests and imports continue to work.

Also updates test patches to match BaseAgent refactoring and
session_context changes from the previous commit.
